2 main important things in this commit: 1 - we separate the store. 2 - we separate the routes. it make the file main.js easier to read and maintain.
